+++ Base32768/Base32768Stream.Write.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.Diagnostics;
+using System.IO;
+using static Kzrnm.Convert.Base32768.Base32768;
+
+namespace Kzrnm.Convert.Base32768
+{
+    /// <summary>
+    /// Stream that encode/decode Base32768.
+    /// </summary>
+    public partial class Base32768Stream : Stream
+    {
+        /// <inheritdoc/>
+        public override void Flush()
+        {
+            throw new NotImplementedException();
+        }
+
+        /// <inheritdoc/>
+        public override void Write(byte[] buffer, int offset, int count)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
